refactor: move IAL 403 redirect into useHouseholdData hook
Moves the 403 + requiredIal detection and redirect logic from
DashboardContent into the useHouseholdData hook so all consumers
get the behavior automatically. Exposes requiresProofing flag
for loading state during redirect.
