+++ src/SEBT.Portal.Web/src/features/household/components/DashboardContent/DashboardContent.tsx
@@ -3,7 +3,6 @@
 import { ApiError } from '@/api'
 import { AnalyticsEvents, useDataLayer } from '@sebt/analytics'
 import { Alert } from '@sebt/design-system'
-import { useRouter } from 'next/navigation'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 
@@ -21,21 +20,8 @@ import { UserProfileCard } from '../UserProfileCard'
 // TODO: Add to CSV: "S2 - Portal Dashboard - Error Heading" and "S2 - Portal Dashboard - Error Description"
 export function DashboardContent() {
   const { t } = useTranslation('dashboard')
-  const router = useRouter()
-  const { data, isLoading, isError, error } = useHouseholdData()
+  const { data, isLoading, isError, error, requiresProofing } = useHouseholdData()
   const { setPageData, setUserData, trackEvent } = useDataLayer()
-
-  // A 403 with requiredIal in the response body means the user's IAL is below
-  // the minimum required by their cases. Redirect to ID proofing.
-  const requiresProofing =
-    error instanceof ApiError &&
-    error.status === 403 &&
-    'requiredIal' in ((error.data as Record<string, unknown>) ?? {})
-  useEffect(() => {
-    if (requiresProofing) {
-      router.push('/login/id-proofing')
-    }
-  }, [requiresProofing, router])
 
   useEffect(() => {
     if (isLoading) return
